refactor of the envelope logic into separate class
The envelope was too annoying to carry around as a tuple and the fact that some projections increase as once goes north
means that most of the time envelopes had to have nasty hacks such as negative 'height'. Re-factor all of the envelope
logic to explicitly name what properties we're using.