fix memory leak
The in-memory temporary datasets were not being unlinked when they were finished with. To make this happen
automagically, the datasets need to be wrapped in an internal _DatasetWrapper class which can take care of freeing
memory after use.

This is a little ugly, perhaps a proxy object would be nicer.